trivial: promote move_vm_ext to folder
In preparation for adding MoveResolverExt
